+++ spec/lib/ember-cli/app_spec.rb
@@ -1,0 +1,9 @@
+require "ember-cli-rails"
+
+describe "User tests Ember app" do
+  it "exits with exit status of 0" do
+    passed = EmberCli["my-app"].test
+
+    expect(passed).to be true
+  end
+end